requirements.txt
python-dotenv==0.20.0
python-telegram-bot==20.20
aiohttp==3.8.3